requirements.txt
# Deep Learning Frameworks
tensorflow>=2.16.0
keras>=3.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
opencv-python>=4.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning Utils
scikit-learn>=1.3.0

# Progress Bars
tqdm>=4.65.0

# Data Augmentation
albumentations>=1.3.0
imgaug>=0.4.0

# Interpretability (optional)
# tf-keras-vis>=0.8.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Web Application (optional)
# flask>=3.0.0
# flask-cors>=4.0.0

# Utilities
h5py>=3.9.0
requests>=2.31.0
